requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
asyncpg==0.29.0
motor==3.4.0
pymongo==4.7.2
psycopg2-binary==2.9.9
scikit-learn==1.4.2
numpy==1.26.4
pandas==2.2.2
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
httpx==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
razorpay==1.4.1
websockets==12.0
